requirements.txt
click==8.1.0
colorama==0.4.4
Flask==2.1.0
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
numpy==1.21.4
opencv-python==4.5.5.64
Pillow==9.0.1
pycodestyle==2.8.0
toml==0.10.2
Werkzeug==2.1.0
joblib==1.1.0
rpi_ws281x==4.3.4
adafruit-circuitpython-neopixel==6.3.0
